% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_advs_details.R
\name{check_adv_available}
\alias{check_adv_available}
\title{Check if advertisement is still available}
\usage{
check_adv_available(html)
}
\arguments{
\item{html}{Xml_document containing the html content of an advertisement
page.}
}
\value{
a boolean.
}
\description{
\code{check_adv_available} checks whether a specific advertisement is still
available on marktplaats.
}
\details{
The function uses the html of a marktplaats advertisement to determine
whether the advertisement is still available or whether it has been removed.
}
\examples{

check_adv_available(html = read_html('http://www.marktplaats.nl/a/telecommunicatie/mobiele-telefoons-apple-iphone/m1106778417-apple-telefoon-4s.html'))

}

