% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_available_advs_overview.R
\name{combine_adv_info_from_page}
\alias{combine_adv_info_from_page}
\title{Combine adv info from a page}
\usage{
combine_adv_info_from_page(titles, urls, filter_admarkt = TRUE)
}
\arguments{
\item{titles}{Vector object containing strings corresponding to advertisement
titles.}

\item{urls}{Vector object containing strings corresponding to advertisement
urls.}

\item{filter_admarkt}{Boolean indicating whether admarkt advertisements need
to be filtered.}
}
\value{
a data.frame object, with columns 'title', 'url' and 'adv_id' and one
  row per advertisement.
}
\description{
\code{combine_adv_info_from_page} combines the titles and urls of
advertisements from a search results page.
}
\details{
The function combines vectors of advertisement titles and urls into a
data.frame object, while also extracting the advertisement ids from the urls
and optionally filtering admarkt advertisements (which are payed
advertisements from companies)
}
\examples{

combine_adv_info_from_page(
   titles = c("Adv A","Adv B"),
   urls = c("https://url_A","https://url_B")
)

}

