% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_advs_images.R
\name{get_adv_images}
\alias{get_adv_images}
\title{Get adv images}
\usage{
get_adv_images(adv_url, storage_dir = "")
}
\arguments{
\item{adv_url}{String containing the url to an advertisement.}

\item{storage_dir}{String indicating the storage location for the downloaded 
images.}
}
\value{
No return.
}
\description{
\code{get_adv_images} downloads and stores all images used in the 
advertisement indicated in the specified url.
}
\details{
The function uses a string containing the url to an advertisement to download
and store all images linked to that advertisement.
}

