% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_adv_titles_from_page}
\alias{get_adv_titles_from_page}
\title{Get the title for each adv on a page}
\usage{
get_adv_titles_from_page(page_html)
}
\arguments{
\item{page_html}{Xml_document containing the html content of a search results
page.}
}
\value{
a vector object containing strings.
}
\description{
\code{get_adv_titles_from_page} gets the title for each of the advertisements on
a search results page.
}
\details{
The function uses the html of a marktplaats search results page to get the
title for each of the advertisments on that page.
}

