% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_advertisement_xml.R
\name{get_advertisement_xml}
\alias{get_advertisement_xml}
\title{Scrape advertisement}
\usage{
get_advertisement_xml(ad_id, wait_time = 0, number_of_tries = 1,
  verbose = F)
}
\arguments{
\item{ad_id}{The ID of a marktplaats advertisement}

\item{wait_time}{To prevent hammering markplaats with rapid requests, you can
specify a delay after collecting the data}

\item{number_of_tries}{If scraping fails we can retry more times}

\item{verbose}{Be chatty about data collection?}
}
\value{
a XML with the HTML of the marktplaats advertisement
}
\description{
Downloads the HTML code of a given advertisement
}

