% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_available_advs_overview.R
\name{get_advs_overview_from_page}
\alias{get_advs_overview_from_page}
\title{Get info on all advs on a page}
\usage{
get_advs_overview_from_page(page_url, filter_admarkt = TRUE, verbose = TRUE)
}
\arguments{
\item{page_url}{String containing the marktplaats url, which corresponds to a
single page with results of a search for advertisements.}

\item{filter_admarkt}{Boolean indicating whether admarkt advertisements need
to be filtered.}

\item{verbose}{Boolean indicating whether information on the progress should
be printed during execution.}
}
\value{
a data.frame object, with columns 'title', 'url' and 'adv_id' and one
  row per advertisement.
}
\description{
\code{get_advs_overview_from_page} gets the info on all advertisments on a
search results page.
}
\details{
The function scrapes a specified marktplaats url corresponding to a search
results page. For each advertisement on the page, the title and url are
scraped and the advertisement id is extracted from the url. The results are
combined into a data.frame and returned.
}
\examples{

get_advs_overview_from_page(
   titles = c("Adv A","Adv B"),
   urls = c("https://url_A","https://url_B")
)

}

