% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_advs_details.R
\name{get_css_element}
\alias{get_css_element}
\title{Get css element from html page}
\usage{
get_css_element(html, css, remove_chars = c(), as_numeric = FALSE,
  expecting_one = TRUE)
}
\arguments{
\item{html}{Xml_document containing the html content of an advertisements
page.}

\item{css}{String indicating the css element to extract.}

\item{remove_chars}{Vector indivating which characters are to be removed from
the result.}

\item{as_numeric}{Boolean indicating whether the result should be returned as
numeric.}

\item{expecting_one}{Boolean indicating whether a single result is expected
or a vector of results.}
}
\value{
a character/numeric or vector object.
}
\description{
\code{get_css_element} gets a specific css element from a html page.
}
\details{
The function uses a css string to extract information from a specific html
node of a html page.
}
\examples{
\dontrun{
get_css_element(
   html = read_html('http://www.marktplaats.nl/z/telecommunicatie/mobiele-telefoons-apple-iphone/iphone.html?query=iphone&categoryId=1953'),
   css = '.price-new',
   expecting_one = F
)
}
}

