% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inactive_merchants.R
\name{get_n_current_advs_of_merchant}
\alias{get_n_current_advs_of_merchant}
\title{Get the number of current advertisements of a merchant}
\usage{
get_n_current_advs_of_merchant(merchant_id, n_tries = 3)
}
\arguments{
\item{merchant_id}{The ID of a marktplaats merchant (a.k.a. counterparty)}

\item{n_tries}{If scraping fails we can retry more times}
}
\value{
a character string
}
\description{
Get the number of current advertisements of a merchant
}
\examples{
\dontrun{
get_n_current_advs_of_merchant(merchant_id = 15818)
}
}
