% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_available_advs_overview.R
\name{get_number_of_adv_pages}
\alias{get_number_of_adv_pages}
\title{Get the number of pages advs}
\usage{
get_number_of_adv_pages(marktplaats_url)
}
\arguments{
\item{marktplaats_url}{String containing the marktplaats url, which
corresponds to the results of a search for advertisements.}
}
\value{
a numeric object.
}
\description{
\code{get_number_of_adv_pages} determines the number of pages with
advertisements.
}
\details{
The function scrapes a specified marktplaats url to determine the number of
results pages with advertisements.
}

