% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_advs_images.R
\name{get_urls_to_adv_images}
\alias{get_urls_to_adv_images}
\title{Get urls to adv images}
\usage{
get_urls_to_adv_images(html)
}
\arguments{
\item{html}{Xml_document containing the html content of an advertisement 
page.}
}
\value{
a vector object containing strings.
}
\description{
\code{get_urls_to_adv_images} gets the url for each of the images used in an
advertisement.
}
\details{
The function uses the html of a marktplaats advertisement to get the 
url for each of images linked to that advertisment.
}
\examples{

get_urls_to_adv_images(html = read_html('http://www.marktplaats.nl/a/telecommunicatie/mobiele-telefoons-apple-iphone/m1106778417-apple-telefoon-4s.html'))

}

