% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_advertisements.R
\name{list_advertisements}
\alias{list_advertisements}
\title{Lists the advertisements listed for a given search url}
\usage{
list_advertisements(url, advertisement_type = c("both", "individuals",
  "companies"), max_pages = Inf)
}
\arguments{
\item{url}{The HTTP URL of a marktplace search page
page.}

\item{advertisement_type}{Choose which type of advertisements to would like to select}

\item{max_pages}{limit the number of pages webscrape. Useful for testing. Default is all pages.}
}
\value{
a data frame with title, url and ID of advertisements
}
\description{
Lists the advertisements listed for a given search url
}
\examples{
\dontrun{
url <- "http://www.marktplaats.nl/z/telecommunicatie/mobiele-telefoons-apple-iphone/iphone.html?query=iphone&categoryId=1953&sortBy=SortIndex"
list_advertisements(url, advertisement_type = "individuals", max_pages = 5)
}
}
