% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ads.R
\name{scrape_ads}
\alias{scrape_ads}
\title{scrape_ads}
\usage{
scrape_ads(ad_ids, ads_per_minute = Inf, report_every_nth_scrape = 10,
  number_of_tries = 2)
}
\arguments{
\item{ad_ids}{vector of ad ids to scrape}

\item{ads_per_minute}{limit the rate of scraping. Default is as fast as possible.}

\item{report_every_nth_scrape}{report scraping every nth ad. Default is 10}

\item{number_of_tries}{Number of tries to scrape an ad if a problem occurs.}
}
\value{
data.frame with features from the ads
}
\description{
scrape_ads
}

