% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_advs_images.R
\name{scrape_adv_images}
\alias{scrape_adv_images}
\title{Scrape adv images}
\usage{
scrape_adv_images(ad_id, storage_dir = "")
}
\arguments{
\item{ad_id}{String containing the url to an advertisement.}

\item{storage_dir}{String indicating the storage location for the downloaded 
images.}
}
\value{
No return.
}
\description{
\code{scrape_adv_images} downloads and stores all images used in the 
advertisement indicated in the specified url.
}
\details{
The function uses a string containing the url to an advertisement to download
and store all images linked to that advertisement.
}
\examples{
\dontrun{
scrape_adv_images(ad_id = "m1016608065", storage_dir = "C:/")
}
}
