% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_advertisement.R
\name{scrape_advertisement}
\alias{scrape_advertisement}
\title{Scrape advertisement}
\usage{
scrape_advertisement(ad_id, verbose = F)
}
\arguments{
\item{id}{The ID of a marktplaats advertisement}
}
\value{
a data.frame with features collected from the ad
}
\description{
Collects a set of features from a given advertisement id
}

