% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_advertisement_images.R
\name{scrape_advertisement_images}
\alias{scrape_advertisement_images}
\title{Scrape advertisement images}
\usage{
scrape_advertisement_images(ad_id, wait_time = 0, number_of_tries = 1,
  verbose = F, include_hash = F)
}
\arguments{
\item{ad_id}{The ID of a marktplaats advertisement}

\item{wait_time}{To prevent hammering markplaats with rapid requests, you can
specify a delay after collecting the data}

\item{number_of_tries}{If scraping fails we can retry more times}

\item{verbose}{Be chatty about data collection?}

\item{include_hash}{Use pHash algoritm to hash image}
}
\value{
a dataframe
}
\description{
returns a data frame with image properties
}

