% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_advertisement.R
\name{scrape_advertisement}
\alias{scrape_advertisement}
\title{Scrape advertisement}
\usage{
scrape_advertisement(ad_id, wait_time = 0, number_of_tries = 1,
  verbose = F)
}
\arguments{
\item{ad_id}{The ID of a marktplaats advertisement}

\item{wait_time}{To prevent hammering markplaats with rapid requests, you can
specify a delay after collecting the data}

\item{number_of_tries}{If scraping fails we can retry more times}

\item{verbose}{Be chatty about data collection?}
}
\value{
a data.frame with features collected from the ad
}
\description{
Collects a set of features from a given advertisement id
}
